+++ modules/graph-layers/src/graph/create-graph-from-data.ts
@@ -4,12 +4,11 @@
 
 import type {Graph, GraphProps} from './graph';
 import type {GraphData} from '../graph-data/graph-data';
-import type {ColumnarGraphColumns} from '../graph-data/columnar-graph-data-builder';
 import {type ArrowGraphData, isArrowGraphData} from '../graph-data/arrow-graph-data';
 import {ArrowGraph} from './arrow-graph';
 import {createTabularGraphFromData} from './create-tabular-graph-from-data';
 
-export function createGraphFromData(data: GraphData | ColumnarGraphColumns | ArrowGraphData, props: GraphProps = {}): Graph {
+export function createGraphFromData(data: GraphData | ArrowGraphData, props: GraphProps = {}): Graph {
   if (isArrowGraphData(data)) {
     return new ArrowGraph(data, props);
   }